+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.8.1 - Unreleased
 
+### Added
+
+New convenience arguments to `.cli`
+
 
 ## Version 0.8.0 - Released 2024-08-14